requirements.txt
# Data generation
faker==24.2.0
pandas==2.2.2
numpy==1.26.4

# Warehouse connectors
snowflake-connector-python==3.10.1
duckdb==0.10.2

# dbt
dbt-core==1.8.7
dbt-snowflake==1.8.4
dbt-duckdb==1.8.1

# AI
anthropic==0.28.0

# Dashboard
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0

# PDF export
weasyprint==62.3

# Utilities
python-dotenv==1.0.1
sqlalchemy==2.0.30
pyarrow==16.0.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0
